+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Master
+
+- Fix OTMVPlans typing
+
 # v0.4.1 (2020-03-19)
 
 - NM B2B 23.5.0